+++ Software.h
@@ -1,4 +1,6 @@
 #pragma once
 #include "Hardware.h"
 #include "Interface.h"
+#include "Memory.h"
+#include "Process.h"
 #include "Game.h"