+++ changelog.md
@@ -1,6 +1,14 @@
 # CPP-Framework
 C++ Framework to build easy application and hack.
 # ChangeLogs 
+## Version 0.1.1
+### Software Memory
+-Software/Memory/Module
+-Software/Memory/VirtualMethodTable
+### Software Main (Windows Process)
+-Software/Process
+
+
 ## Version 0.1
 - Main Base Version
 ### Software Interfaces